Preserve semantic projection ownership
