+++ OfficeIMO.Html/Conversion/HtmlEditableLayoutProjection.cs
@@ -75,7 +75,7 @@ public sealed class HtmlEditableLayoutProjection {
 }
 
 /// <summary>Creates one shared editable-layout plan for DOCX, RTF, XLSX, and PPTX adapters.</summary>
-public static class HtmlEditableLayoutProjector {
+public static partial class HtmlEditableLayoutProjector {
     internal const string RegionAttribute = "data-officeimo-editable-layout-region";
     internal const string ImageAttribute = "data-officeimo-editable-layout-image";
     private const string ImageSourcePrefix = "img[officeimo-layout-image=";
@@ -160,6 +160,7 @@ public static class HtmlEditableLayoutProjector {
         var multiChildLayoutKeys = new HashSet<string>(StringComparer.Ordinal);
         var nestedLayoutPlacementKeys = new HashSet<string>(StringComparer.Ordinal);
         var bookmarkTargetCandidates = new List<IElement>();
+        var commentBearingCandidates = new List<IElement>();
         var effectCandidates = new List<(IElement Element, string Detail)>();
         var mixedInlineImageCandidates = new List<IElement>();
         foreach (IElement element in adapterDocument.QuerySelectorAll("*")) {
@@ -167,6 +168,11 @@ public static class HtmlEditableLayoutProjector {
                 || !TryGetCandidateKind(element, style, out HtmlEditableLayoutRegionKinds candidateKind)
                 || (regionKinds & candidateKind) == 0) continue;
             if (semanticFlowRoots.Any(root => root.Contains(element))) continue;
+            if (ContainsHtmlComment(element)) {
+                commentBearingCandidates.Add(element);
+                semanticFlowRoots.Add(element);
+                continue;
+            }
             if (ContainsBookmarkTarget(element)) {
                 bookmarkTargetCandidates.Add(element);
                 semanticFlowRoots.Add(element);
@@ -223,6 +229,12 @@ public static class HtmlEditableLayoutProjector {
         }
 
         var diagnostics = new HtmlDiagnosticReport();
+        foreach (IElement element in commentBearingCandidates) {
+            diagnostics.Add("OfficeIMO.Html", HtmlEditableLayoutDiagnosticCodes.PlacementSimplified,
+                "An editable layout region stayed in semantic flow so its raw HTML comments could remain available to the destination.",
+                HtmlDiagnosticSeverity.Warning, HtmlRenderStyleResolver.DescribeSource(element),
+                "htmlComment=true; semanticFlow=true", OfficeConversionLossKind.Approximation);
+        }
         foreach (IElement element in bookmarkTargetCandidates) {
             diagnostics.Add("OfficeIMO.Html", HtmlEditableLayoutDiagnosticCodes.PlacementSimplified,
                 "An editable layout region stayed in semantic flow so its bookmark target would remain addressable.",
@@ -344,6 +356,9 @@ public static class HtmlEditableLayoutProjector {
         }
 
         var acceptedKeys = new HashSet<string>(accepted.Select(region => region.SourceKey), StringComparer.Ordinal);
+        var semanticSectionOwnerKeys = new HashSet<string>(accepted
+            .Where(region => IsSemanticSectionOwner(candidateElements[region.SourceKey], adapterDocument))
+            .Select(region => region.SourceKey), StringComparer.Ordinal);
         IReadOnlyDictionary<string, IReadOnlyList<IHtmlImageElement>> orderedSourceImages = accepted.ToDictionary(
             region => region.SourceKey,
             region => CreateOrderedSourceImages(region, candidateElements[region.SourceKey], styles),
@@ -372,8 +387,16 @@ public static class HtmlEditableLayoutProjector {
         }
         foreach (IElement element in remainingDocument.QuerySelectorAll("[" + RegionAttribute + "]").ToArray()) {
             string? sourceKey = element.GetAttribute(RegionAttribute);
-            if (sourceKey != null && acceptedKeys.Contains(sourceKey)) element.Remove();
-            else element.RemoveAttribute(RegionAttribute);
+            if (sourceKey != null && acceptedKeys.Contains(sourceKey)) {
+                if (semanticSectionOwnerKeys.Contains(sourceKey)) {
+                    foreach (INode child in element.ChildNodes.ToArray()) element.RemoveChild(child);
+                    element.RemoveAttribute(RegionAttribute);
+                } else {
+                    element.Remove();
+                }
+            } else {
+                element.RemoveAttribute(RegionAttribute);
+            }
         }
         foreach (IElement image in remainingDocument.QuerySelectorAll("[" + ImageAttribute + "]").ToArray()) {
             image.RemoveAttribute(ImageAttribute);
@@ -582,181 +605,6 @@ public static class HtmlEditableLayoutProjector {
             if (ReferenceEquals(owners[index], element) || owners[index].Contains(element)) return index + 1;
         }
         return 0;
-    }
-
-    private static bool ContainsSemanticRichContent(
-        IElement element,
-        IReadOnlyDictionary<IElement, HtmlComputedStyle> styles) {
-        if (string.Equals(element.LocalName, "img", StringComparison.OrdinalIgnoreCase)) return false;
-        if (SemanticRichElementNames.Contains(element.LocalName)) return true;
-        if (HasDistinctRichTextStyle(element, styles)) return true;
-        if (HasVisibleRootBorder(element, styles)) return true;
-        return element.QuerySelectorAll("*").Any(child => {
-            if (string.Equals(child.LocalName, "img", StringComparison.OrdinalIgnoreCase)) return false;
-            return SemanticRichElementNames.Contains(child.LocalName)
-                || HasDistinctRichTextStyle(child, styles)
-                || HasDistinctStyle(child, styles, RichDescendantVisualStyleProperties);
-        });
-    }
-
-    private static bool ContainsBookmarkTarget(IElement element) {
-        if (!string.IsNullOrWhiteSpace(element.GetAttribute("id"))) return true;
-        return element.QuerySelectorAll("[id]").Any(target =>
-            !string.IsNullOrWhiteSpace(target.GetAttribute("id")));
-    }
-
-    private static bool ContainsMixedInlineImageContent(
-        IElement element,
-        IReadOnlyDictionary<IElement, HtmlComputedStyle> styles) {
-        var contentOrder = new List<bool>();
-        AppendInlineContentOrder(element, element, styles, contentOrder);
-        return contentOrder.Contains(true) && contentOrder.Contains(false);
-    }
-
-    private static bool HasMultipleVisibleLayoutChildren(
-        IElement element,
-        IReadOnlyDictionary<IElement, HtmlComputedStyle> styles) {
-        int visibleChildren = 0;
-        foreach (INode child in element.ChildNodes) {
-            if (child is IElement childElement) {
-                if (!IsProjectionElementVisible(childElement, element, styles)
-                    || (string.IsNullOrWhiteSpace(childElement.TextContent)
-                        && childElement.QuerySelector("img") == null)) continue;
-            } else if (string.IsNullOrWhiteSpace(child.TextContent)) {
-                continue;
-            }
-            visibleChildren++;
-            if (visibleChildren > 1) return true;
-        }
-        return false;
-    }
-
-    private static bool IsFlexOrGridDisplay(HtmlComputedStyle style) {
-        string display = style.GetValue("display").Trim();
-        return display.Equals("flex", StringComparison.OrdinalIgnoreCase)
-            || display.Equals("inline-flex", StringComparison.OrdinalIgnoreCase)
-            || display.Equals("grid", StringComparison.OrdinalIgnoreCase)
-            || display.Equals("inline-grid", StringComparison.OrdinalIgnoreCase);
-    }
-
-    private static bool ContainsNestedLayoutPlacement(
-        IElement element,
-        IReadOnlyDictionary<IElement, HtmlComputedStyle> styles) {
-        return element.QuerySelectorAll("*").Any(child =>
-            child is not IHtmlImageElement
-            && IsProjectionElementVisible(child, element, styles)
-            && styles.TryGetValue(child, out HtmlComputedStyle? childStyle)
-            && HasLayoutPlacementStyle(childStyle));
-    }
-
-    private static bool HasLayoutPlacementStyle(HtmlComputedStyle style) {
-        string position = style.GetValue("position").Trim();
-        string floatSide = style.GetValue("float").Trim();
-        return position.Equals("absolute", StringComparison.OrdinalIgnoreCase)
-            || position.Equals("fixed", StringComparison.OrdinalIgnoreCase)
-            || floatSide.Equals("left", StringComparison.OrdinalIgnoreCase)
-            || floatSide.Equals("right", StringComparison.OrdinalIgnoreCase)
-            || IsFlexOrGridDisplay(style);
-    }
-
-    private static void AppendInlineContentOrder(
-        INode node,
-        IElement region,
-        IReadOnlyDictionary<IElement, HtmlComputedStyle> styles,
-        ICollection<bool> contentOrder) {
-        foreach (INode child in node.ChildNodes) {
-            if (child is IHtmlImageElement image) {
-                if (IsProjectionImageVisible(image, region, styles)) contentOrder.Add(true);
-            } else if (child is IElement childElement) {
-                AppendInlineContentOrder(childElement, region, styles, contentOrder);
-            } else if (!string.IsNullOrWhiteSpace(child.TextContent)) {
-                contentOrder.Add(false);
-            }
-        }
-    }
-
-    private static bool TryGetNonNativeRegionEffect(
-        IElement element,
-        IReadOnlyDictionary<IElement, HtmlComputedStyle> styles,
-        out string detail) {
-        var effects = new List<string>();
-        for (IElement? current = element; current != null; current = current.ParentElement) {
-            AddNonNativeEffects(current, styles, effects, "");
-        }
-        foreach (IElement descendant in element.QuerySelectorAll("*")) {
-            AddNonNativeEffects(descendant, styles, effects, "descendant:");
-        }
-        detail = string.Join("; ", effects.Distinct(StringComparer.OrdinalIgnoreCase));
-        return effects.Count > 0;
-    }
-
-    private static void AddNonNativeEffects(
-        IElement element,
-        IReadOnlyDictionary<IElement, HtmlComputedStyle> styles,
-        ICollection<string> effects,
-        string prefix) {
-        if (!styles.TryGetValue(element, out HtmlComputedStyle? style)) return;
-        string opacity = style.GetValue("opacity").Trim();
-        if (element is not IHtmlImageElement
-            && double.TryParse(opacity, System.Globalization.NumberStyles.Float,
-                System.Globalization.CultureInfo.InvariantCulture, out double parsedOpacity)
-            && parsedOpacity < 0.999D) {
-            effects.Add(prefix + "opacity=" + opacity);
-        }
-        AddNonDefaultEffect(style, "transform", "none", effects, prefix);
-        AddNonDefaultEffect(style, "clip-path", "none", effects, prefix);
-        AddNonDefaultEffect(style, "filter", "none", effects, prefix);
-        AddNonDefaultEffect(style, "mix-blend-mode", "normal", effects, prefix);
-        AddNonDefaultEffect(style, "overflow", "visible", effects, prefix);
-        AddNonDefaultEffect(style, "overflow-x", "visible", effects, prefix);
-        AddNonDefaultEffect(style, "overflow-y", "visible", effects, prefix);
-    }
-
-    private static void AddNonDefaultEffect(
-        HtmlComputedStyle style,
-        string property,
-        string defaultValue,
-        ICollection<string> effects,
-        string prefix = "") {
-        string value = style.GetValue(property).Trim();
-        if (value.Length > 0 && !value.Equals(defaultValue, StringComparison.OrdinalIgnoreCase)) {
-            effects.Add(prefix + property + "=" + value);
-        }
-    }
-
-    private static bool HasDistinctRichTextStyle(
-        IElement element,
-        IReadOnlyDictionary<IElement, HtmlComputedStyle> styles) {
-        return HasDistinctStyle(element, styles, RichTextStyleProperties);
-    }
-
-    private static bool HasInheritedRichTextStyle(
-        IElement element,
-        IReadOnlyDictionary<IElement, HtmlComputedStyle> styles) {
-        return styles.TryGetValue(element, out HtmlComputedStyle? style)
-            && RichTextStyleProperties.Any(property => style.IsInheritedValue(property)
-                && !string.IsNullOrWhiteSpace(style.GetValue(property)));
-    }
-
-    private static bool HasVisibleRootBorder(
-        IElement element,
-        IReadOnlyDictionary<IElement, HtmlComputedStyle> styles) {
-        return styles.TryGetValue(element, out HtmlComputedStyle? style)
-            && HtmlCssBoxStrokeParser.HasBorderDeclaration(style);
-    }
-
-    private static bool HasDistinctStyle(
-        IElement element,
-        IReadOnlyDictionary<IElement, HtmlComputedStyle> styles,
-        IReadOnlyList<string> properties) {
-        if (!styles.TryGetValue(element, out HtmlComputedStyle? style)
-            || element.ParentElement == null
-            || !styles.TryGetValue(element.ParentElement, out HtmlComputedStyle? parentStyle)) return false;
-        return properties.Any(property => !style.IsInheritedValue(property)
-            && !style.IsResetValue(property)
-            && !string.IsNullOrWhiteSpace(style.GetValue(property))
-            && !string.Equals(style.GetValue(property), parentStyle.GetValue(property),
-                StringComparison.OrdinalIgnoreCase));
     }
 
     private static bool DiagnosticsAreEquivalent(HtmlDiagnostic first, HtmlDiagnostic second) =>
